extracted all private classes into their own files
