+++ web-ui/src/main/java/org/vaadin/example/RemoveWhereLayoutRequestEvent.java
@@ -1,0 +1,9 @@
+package org.vaadin.example;
+import com.vaadin.flow.component.ComponentEvent;
+
+public class RemoveWhereLayoutRequestEvent extends ComponentEvent<WhereTextBoxesLayout> {
+    public RemoveWhereLayoutRequestEvent(WhereTextBoxesLayout source, boolean fromClient) {
+        super(source, fromClient);
+        System.out.println("removeWhereLayoutRequestEvent Fired");
+    }
+}
