+++ web-ui/src/main/java/org/vaadin/example/ResultUpdateEvent.java
@@ -1,0 +1,9 @@
+package org.vaadin.example;
+import com.vaadin.flow.component.ComponentEvent;
+
+public class ResultUpdateEvent extends ComponentEvent<QueryResults> {
+    public ResultUpdateEvent(QueryResults source, boolean fromClient) {
+        super(source, fromClient);
+        System.out.println("Result Fired");
+    }
+}
