+++ web-ui/src/main/java/org/vaadin/example/SideBarLayout.java
@@ -1,0 +1,8 @@
+package org.vaadin.example;
+import com.vaadin.flow.component.orderedlayout.VerticalLayout;
+
+public class SideBarLayout extends VerticalLayout {
+    SideBarLayout() {
+        getStyle().set("background-color", "#3458eb");
+    }
+}
